third_party/WebKit/LayoutTests/webaudio/AudioParam/worklet-warnings-expected.txt

CONSOLE WARNING: line 36: AudioParam value setter will become equivalent to AudioParam.setValueAtTime() in M65, around March 2018  See https://webaudio.github.io/web-audio-api/#dom-audioparam-value for more details.
CONSOLE WARNING: line 39: AudioWorklet("noise-generator").amplitude.value 99 outside nominal range [0, 1]; value will be clamped.
CONSOLE WARNING: line 43: AudioWorklet("noise-generator").amplitude.setValueAtTime value -1 outside nominal range [0, 1]; value will be clamped.
CONSOLE WARNING: line 44: AudioWorklet("noise-generator").amplitude.linearRampToValueAtTime value 5 outside nominal range [0, 1]; value will be clamped.
This is a testharness.js-based test.
PASS # AUDIT TASK RUNNER STARTED.
PASS > [AudioWorklet AudioParam] Generate warnings if outside nominal range
PASS   Rendering succeeded is true.
PASS < [AudioWorklet AudioParam] All assertions passed. (total 1 assertions)
PASS # AUDIT TASK RUNNER FINISHED: 1 tasks ran successfully.
Harness: the test ran to completion.
